requirements.txt
# Optimized requirements for cloud deployment
# Core automation packages
schedule==1.2.0
python-dotenv==1.0.0

# Web framework and authentication
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1

# YouTube API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# HTTP requests
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Cloud deployment
gunicorn==21.2.0

# Image processing (lightweight)
Pillow==10.0.1

# Basic text processing
Jinja2==3.1.2

# YAML for config
PyYAML==6.0.1
